advertisment-manager/requirements.txt
asgiref==3.5.2
boto3==1.26.3
botocore==1.29.3
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-rest-framework==0.1.0
django-storages==1.13.1
djangorestframework==3.14.0
idna==3.4
jmespath==1.0.1
mysqlclient==2.1.1
pika==1.3.1
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
